+++ src/thinger/system/backup.h
@@ -4,8 +4,11 @@ class ThingerMonitorBackup {
 protected:
     ThingerMonitorConfig config_;
     std::string name_;
+    std::string tag_;
 
-    ThingerMonitorBackup(ThingerMonitorConfig& config, const std::string& name) : config_(config), name_(name) {}
+    ThingerMonitorBackup(ThingerMonitorConfig& config, const std::string& name, const std::string& tag)
+      : config_(config), name_(name), tag_(tag)
+    {}
 
 public: